Move tests related to async coros to their own module
