+++ tests/test_cases/test_cocotb/Makefile
@@ -41,6 +41,7 @@ MODULE := "\
 	test_scheduler,\
 	test_clock,\
 	test_edge_triggers,\
+	test_async_coroutines,\
 	"
 
 ifeq ($(shell python -c "import sys; print(sys.version_info >= (3, 6))"), "True")
